// ui/src/lib/api-client/models/GetAuthIdentityProviders200ResponseInner.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Proxy Smart
 * SMART on FHIR Proxy + Healthcare Administration API using Keycloak and Elysia
 *
 * The version of the OpenAPI document: 0.0.1-alpha.202508151735.a205e19
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface GetAuthIdentityProviders200ResponseInner
 */
export interface GetAuthIdentityProviders200ResponseInner {
    /**
     * Provider alias
     * @type {string}
     * @memberof GetAuthIdentityProviders200ResponseInner
     */
    alias: string;
    /**
     * Provider type
     * @type {string}
     * @memberof GetAuthIdentityProviders200ResponseInner
     */
    providerId: string;
    /**
     * Display name
     * @type {string}
     * @memberof GetAuthIdentityProviders200ResponseInner
     */
    displayName: string;
    /**
     * Whether provider is enabled
     * @type {boolean}
     * @memberof GetAuthIdentityProviders200ResponseInner
     */
    enabled: boolean;
}

/**
 * Check if a given object implements the GetAuthIdentityProviders200ResponseInner interface.
 */
export function instanceOfGetAuthIdentityProviders200ResponseInner(value: object): value is GetAuthIdentityProviders200ResponseInner {
    if (!('alias' in value) || value['alias'] === undefined) return false;
    if (!('providerId' in value) || value['providerId'] === undefined) return false;
    if (!('displayName' in value) || value['displayName'] === undefined) return false;
    if (!('enabled' in value) || value['enabled'] === undefined) return false;
    return true;
}

export function GetAuthIdentityProviders200ResponseInnerFromJSON(json: any): GetAuthIdentityProviders200ResponseInner {
    return GetAuthIdentityProviders200ResponseInnerFromJSONTyped(json, false);
}

export function GetAuthIdentityProviders200ResponseInnerFromJSONTyped(json: any, ignoreDiscriminator: boolean): GetAuthIdentityProviders200ResponseInner {
    if (json == null) {
        return json;
    }
    return {
        
        'alias': json['alias'],
        'providerId': json['providerId'],
        'displayName': json['displayName'],
        'enabled': json['enabled'],
    };
}

export function GetAuthIdentityProviders200ResponseInnerToJSON(json: any): GetAuthIdentityProviders200ResponseInner {
    return GetAuthIdentityProviders200ResponseInnerToJSONTyped(json, false);
}

export function GetAuthIdentityProviders200ResponseInnerToJSONTyped(value?: GetAuthIdentityProviders200ResponseInner | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'alias': value['alias'],
        'providerId': value['providerId'],
        'displayName': value['displayName'],
        'enabled': value['enabled'],
    };
}